requirements.txt
# Azure & OpenAI
openai>=1.12.0
azure-ai-projects>=1.0.0b10
azure-identity>=1.15.0
azure-ai-inference>=1.0.0b5

# Salesforce
simple-salesforce>=1.12.0

# Environment & Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0

# Web (Future)
flask>=3.0.0
flask-cors>=4.0.0